requirements/dev.txt
# requirements/dev.txt requisitos para entorno de desarrollo
-r common.txt
django-debug-toolbar==0.9.4
